+++ Editor/SaintsEditor.cs
@@ -357,8 +357,8 @@ namespace SaintsField.Editor
                             int order = orderProp?.Order ?? int.MinValue;
 
                             // inspector does not care about inherited/new method. It just needs to use the last one
-                            fieldWithInfos.RemoveAll(each => each.RenderType == SaintsRenderType.Method && each.MethodInfo.Name == methodInfo.Name);
-                            methodInfos.RemoveAll(each => each.RenderType == SaintsRenderType.Method && each.MethodInfo.Name == methodInfo.Name);
+                            fieldWithInfos.RemoveAll(each => each.InherentDepth < inherentDepth && each.RenderType == SaintsRenderType.Method && each.MethodInfo.Name == methodInfo.Name);
+                            methodInfos.RemoveAll(each => each.InherentDepth < inherentDepth && each.RenderType == SaintsRenderType.Method && each.MethodInfo.Name == methodInfo.Name);
 
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_SAINTS_EDITOR_METHOD
                                 Debug.Log($"[{systemType}] method: {methodInfo.Name}");